Test roundtrip with a binary cursor for both geometry and geography
With this commit the binary.sql testcase doesn't test the actual
binary output anymore but rather uses canonical binary output to
copy objects to a temporary file, then uses canonical binary input
to read the objects again into a new table and then compares the
original and the final tables.

In order for this to work run_test was modified to pass a :tmpfile
variable to testcases. Next stop: typmod for canonical binary input.

See ticket #850 for more info

git-svn-id: http://svn.osgeo.org/postgis/trunk@8861 b70326c6-7e19-0410-871a-916f4a2858ee